Move fees section to dedicated component
